Add files via upload
2. (3 pts) A chocolate cookie recipe calls for the following ingredients:
● 2.5 cups of sugar
● 1.5 cups of butter
● 4 cups of flour
● 0.8 cup of chocolate chips
The recipe produces 60 chocolate cookies with this amount of the ingredients.
Write a complete program in a cpp file that asks the user how many cookies he or she
wants to make, and then displays the number of cups of each ingredient needed for the
specified number of cookies

3. (3 pts) Write a complete program in a cpp file that calculates the average rainfall for four
months. The program should have reasonable prompts asking the user to enter the name
of each month, such as December or January, and the amount of rain (in inches) that fell
each month. The program should display a message similar to the following:
The average rainfall for December, January, February, and March is 8.91 inches.